@@ -2,6 +2,7 @@
 
 namespace Externals\Application\Controller;
 
+use Bref\Framework\Controller;
 use Externals\User\User;
 use Externals\User\UserRepository;
 use League\OAuth2\Client\Provider\AbstractProvider;
@@ -9,56 +10,45 @@ use League\OAuth2\Client\Provider\Exception\IdentityProviderException;
 use League\OAuth2\Client\Provider\GithubResourceOwner;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use PSR7Session\Http\SessionMiddleware;
+use PSR7Sessions\Storageless\Http\SessionMiddleware;
+use PSR7Sessions\Storageless\Session\SessionInterface;
 use Twig\Environment;
-use Zend\Diactoros\Response\HtmlResponse;
-use Zend\Diactoros\Response\RedirectResponse;
 
-class UserController
+class LoginController extends Controller
 {
-    /** @var UserRepository */
-    private $userRepository;
-
-    /** @var Environment */
-    private $twig;
-
-    /** @var AbstractProvider */
-    private $authProvider;
-
-    public function __construct(UserRepository $userRepository, Environment $twig, AbstractProvider $authProvider)
-    {
-        $this->userRepository = $userRepository;
-        $this->twig = $twig;
-        $this->authProvider = $authProvider;
+    public function __construct(
+        private UserRepository $userRepository,
+        private Environment $twig,
+        private AbstractProvider $authProvider
+    ) {
     }
 
-    public function login(ServerRequestInterface $request): ResponseInterface
+    public function handle(ServerRequestInterface $request): ResponseInterface
     {
-        newrelic_name_transaction('login');
-
         $session = $request->getAttribute(SessionMiddleware::SESSION_ATTRIBUTE);
+        assert($session instanceof SessionInterface);
 
         // If already logged in
         $user = $session->get('user');
         if ($user instanceof User) {
-            return new RedirectResponse('/');
+            return $this->redirectResponse('/');
         }
 
-        $code = $request->getQueryParams()['code'] ?? null;
+        $code = $this->queryParameter($request, 'code');
         if (! $code) {
             // Get an authorization code from GitHub
             $redirectUrl = $this->authProvider->getAuthorizationUrl();
             $state = $this->authProvider->getState();
             $session->set('oauth2state', $state);
-            return new RedirectResponse($redirectUrl);
+            return $this->redirectResponse($redirectUrl);
         }
 
         // Check given state against previously stored one to mitigate CSRF attack
-        $requestOauthState = $request->getQueryParams()['state'] ?? null;
+        $requestOauthState = $this->queryParameter($request, 'state');
         $sessionOauthState = $session->get('oauth2state');
         if (empty($requestOauthState) || ($requestOauthState !== $sessionOauthState)) {
             $session->remove('oauth2state');
-            return new HtmlResponse($this->twig->render('auth/login-error.html.twig', [
+            return $this->htmlResponse($this->twig->render('auth/login-error.html.twig', [
                 'error' => 'Invalid state',
             ]), 400);
         }
@@ -71,7 +61,7 @@ class UserController
             /** @var GithubResourceOwner $userProfile */
             $userProfile = $this->authProvider->getResourceOwner($accessToken);
         } catch (IdentityProviderException $e) {
-            return new HtmlResponse($this->twig->render('auth/login-error.html.twig', [
+            return $this->htmlResponse($this->twig->render('auth/login-error.html.twig', [
                 'error' => $e->getMessage(),
             ]), 400);
         }
@@ -79,14 +69,6 @@ class UserController
         $user = $this->userRepository->getOrCreate((string) $userProfile->getId(), (string) $userProfile->getNickname());
         $session->set('user', $user);
 
-        return new RedirectResponse('/');
-    }
-
-    public function logout(ServerRequestInterface $request): RedirectResponse
-    {
-        newrelic_name_transaction('logout');
-        $session = $request->getAttribute(SessionMiddleware::SESSION_ATTRIBUTE);
-        $session->remove('user');
-        return new RedirectResponse('/');
+        return $this->redirectResponse('/');
     }
 }
