@@ -8,11 +8,11 @@ use Externals\Email\Email;
 
 class RssBuilder
 {
-    /** @var string $host The base url to use for links. */
-    private $host;
+    /** The base url to use for links. */
+    private string $host;
 
-    /** @var DomDocument|null $dom The current xml document being built. */
-    private $dom;
+    /** The current xml document being built. */
+    private ?DomDocument $dom;
 
     public function __construct(string $host)
     {
