@@ -3,21 +3,26 @@
 namespace Externals\Application\Controller;
 
 use Bref\Framework\Controller;
+use Externals\Email\EmailRepository;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Twig\Environment;
 
-class NotFoundController extends Controller
+class TopController extends Controller
 {
     public function __construct(
         private Environment $twig,
+        private EmailRepository $repository
     ) {
     }
 
     public function handle(ServerRequestInterface $request): ResponseInterface
     {
-        return $this->htmlResponse($this->twig->render('404.html.twig', [
-            'user' => $request->getAttribute('user'),
-        ]), 404);
+        $user = $request->getAttribute('user');
+
+        return $this->htmlResponse($this->twig->render('top.html.twig', [
+            'threads' => $this->repository->findTopThreads(1, $user),
+            'user' => $user,
+        ]));
     }
 }
